Bugfix log1p for non-float dense arrays

+++ scanpy/tests/helpers.py
@@ -19,7 +19,7 @@ def check_rep_mutation(func, X, **kwargs):
         X=X.copy(),
         layers={"layer": X.copy()},
         obsm={"obsm": X.copy()},
-        dtype=np.float64,
+        dtype=X.dtype,
     )
     adata_X = func(adata, copy=True, **kwargs)
     adata_layer = func(adata, layer="layer", copy=True, **kwargs)
